requirements.txt
discord.py==2.4.0
FlightRadar24==0.3.1
